Minor typographic/formatting edits in docs/examples (#105)
* docs: Render 1e-1 better

Previously the rendering of 1e-1 made it look like "one multiplied
by e minus 1". Adding the braces removes the spaces around the '-'
sign.

* example: Render quanity and units consistently

Through the example document the quantity and units are not rendered
in 'math mode', except for '2000 N'. This change removes the 'math
mode' from the 2000 N to be consistent.

* example: Remove hyperlink from code block

Hyperlinks do not display correctly when formatted a code
block. Instead keep the code block and add the hyperlink immediately
afterwards.

* example: Correct typographic error