added warning supressions
When using multiple custom Sphinx objects (one per builder, i.e. one each for linkcheck and html), a bunch of warnings get thrown when creating the second object due to all of the nodes/roles/directives already existing from the first object. There's not a good way around this in Sphinx, but I can't think of an instance where these warnings would be called in normal docs building, so we can suppress them for now to get this working in the CI